weather view

+++ app/src/main/java/com/weatherapp/models/Sys.kt
@@ -10,7 +10,7 @@ data class Sys(
     @SerializedName("message")
     val message: Double,
     @SerializedName("sunrise")
-    val sunrise: Int,
+    val sunrise: Long,
     @SerializedName("sunset")
-    val sunset: Int
+    val sunset: Long
 ): Serializable
